aktualizr: Run in a systemd timer in "run-once" mode
Currently aktualizr polling does not work with systemd automounts:
it eventually times out trying to scan the NFS mount point and then is
locked up and the system needs to be rebooted. JIRA ticket TOR-2647 covers
this issue and when resolved we be able to remove this code.

Instead of standard polling we will create a systemd timer to check
if the mountpoint exists and if the metadata directory is present. if
so then our timer will run aktualir in "run-once" mode.

Signed-off-by: Drew Moseley <drew@moseleynet.net>